ai/requirements.txt
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
spacy>=3.7.0
networkx>=3.1
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
